Fix: revert part of commit (#2143)
revert part of https://github.com/greenbone/openvas-scanner/commit/13c2f987ad7b49ab8a6260cf7ed350636ad10a9a, because it doesn't compiles on i386 and armhf

Close #2140

+++ nasl/nasl_misc_funcs.c
@@ -817,7 +817,7 @@ nasl_gunzip (lex_ctxt *lexic)
 {
   tree_cell *retc;
   void *data, *uncompressed;
-  size_t datalen, uncomplen;
+  unsigned long datalen, uncomplen;
 
   data = get_str_var_by_name (lexic, "data");
   if (data == NULL)
@@ -842,7 +842,7 @@ nasl_gzip (lex_ctxt *lexic)
 {
   tree_cell *retc;
   void *data, *compressed, *headerformat;
-  size_t datalen, complen;
+  unsigned long datalen, complen;
 
   data = get_str_var_by_name (lexic, "data");
   if (data == NULL)
